requirements.txt
asgiref==3.2.10
bidict==0.21.2
cffi==1.14.6
Django==3.0.2
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.8.0
dnspython==1.16.0
eventlet==0.31.1
gevent==21.8.0
greenlet==1.1.1
gunicorn==20.1.0
importlib-metadata==4.6.4
Markdown==3.3.4
npm==0.1.1
optional-django==0.1.0
Pillow==8.3.1
pycparser==2.20
PyJWT==2.1.0
python-engineio==4.2.1
python-socketio==5.4.0
pytz==2020.1
six==1.16.0
sqlparse==0.3.1
typing-extensions==3.10.0.0
zipp==3.5.0
zope.event==4.5.0
zope.interface==5.4.0
